Remove redundant calls to check_argument_units_
check_argument_units_ was being called in several functions before
either a call to convert_angle_ or another function that called
check_argument_units_ or convert_angle.

One could argue that this is acceptable for readability's sake but
for a library where performance is important I believe it's best to
to remove these redundant calls.